# apps/backend/auth-service/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "auth-service",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
...
		"/v1/auth/google": {
			"get": {
				"operationId": "InitiateGoogleSignIn",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"url": {
											"type": "string"
										}
									},
									"required": [
										"url"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Initiate Google Sign-In",
				"summary": "Initiates the Google Sign-In process and redirects to Google's authorization page.",
				"tags": [
					"Google Intergrate AWS Cognito"
				],
				"security": [],
				"parameters": []
			}
		},
		"/v1/auth/google/callback": {
			"get": {
				"operationId": "GoogleCallback",
				"responses": {
					"204": {
						"description": "No content"
					},
					"409": {
						"description": "- Error response in case of a failure.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "- Error response in case of a failure.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Google Callback",
				"summary": "Handles the callback from Google with the authorization code.",
				"tags": [
					"Google Intergrate AWS Cognito"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The authorization code provided by Google.",
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/refresh-token": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "- Error response in case of failure.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Refresh Token",
				"summary": "Exchanges a valid refresh token for a new set of access and ID tokens.",
				"tags": [
					"Google Intergrate AWS Cognito"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}